conditional dns resolution binding
