+++ tests/other_test.go
@@ -2,7 +2,6 @@
 package test
 
 import (
-	"github.com/gruntwork-io/terratest/modules/terraform"
 	"testing"
 
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
@@ -20,27 +19,6 @@ func TestRunBasicExample(t *testing.T) {
 		ResourceGroup: resourceGroup,
 	})
 
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-func TestRunHubAndSpokeDelegatedExample(t *testing.T) {
-	t.Parallel()
-
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:       t,
-		TerraformDir:  hubAndSpokeDelegatedExampleTerraformDir,
-		Prefix:        "has-slz",
-		ResourceGroup: resourceGroup,
-		Region:        "us-south",
-		PostApplyHook: func(options *testhelper.TestOptions) error {
-			terraformOptions := options.TerraformOptions
-			terraformOptions.Vars["update_delegated_resolver"] = true
-			_, err := terraform.ApplyE(options.Testing, terraformOptions)
-			return err
-		},
-	})
 	output, err := options.RunTestConsistency()
 	assert.Nil(t, err, "This should not have errored")
 	assert.NotNil(t, output, "Expected some output")
