+++ tests/pr_test.go
@@ -322,3 +322,25 @@ func TestRunUpgradeFullyConfigurable(t *testing.T) {
 		assert.Nil(t, err, "This should not have errored")
 	}
 }
+
+func TestRunHubAndSpokeDelegatedExample(t *testing.T) {
+	t.Parallel()
+
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
+		Testing:       t,
+		TerraformDir:  hubAndSpokeDelegatedExampleTerraformDir,
+		Prefix:        "has-slz",
+		ResourceGroup: resourceGroup,
+		Region:        "us-south",
+		PostApplyHook: func(options *testhelper.TestOptions) error {
+			terraformOptions := options.TerraformOptions
+			terraformOptions.Vars["update_delegated_resolver"] = true
+			_, err := terraform.ApplyE(options.Testing, terraformOptions)
+			return err
+		},
+	})
+
+	output, err := options.RunTestConsistency()
+	assert.Nil(t, err, "This should not have errored")
+	assert.NotNil(t, output, "Expected some output")
+}
